Factor CustomProvider into its own file within cloud to match structure of Azure, GCP, and AWS; implement ClusterName for Provider interface; use ClusterName instead of Cluster ID for cost data model and aggregate data; clean up and make some cod emore idiomatic.
